--- area_blocks.csv
@@ -1,7 +1,0 @@
-area_block_id,area_block_name
-yokohama,横浜市
-Kawasaki,川崎市
-yokosukamiura,横須賀三浦地区
-kenoh,県央地区
-syonan,湘南地区
-kensei,県西地区
