--- arealist.csv
@@ -1,59 +1,0 @@
-area_id,area_name,area_block
-1,横浜市鶴見区,横浜市
-2,横浜市神奈川区,横浜市
-3,横浜市西区,横浜市
-4,横浜市中区,横浜市
-5,横浜市南区,横浜市
-6,横浜市保土ケ谷区,横浜市
-7,横浜市磯子区,横浜市
-8,横浜市金沢区,横浜市
-9,横浜市港北区,横浜市
-10,横浜市戸塚区,横浜市
-11,横浜市港南区,横浜市
-12,横浜市旭区,横浜市
-13,横浜市緑区,横浜市
-14,横浜市瀬谷区,横浜市
-15,横浜市栄区,横浜市
-16,横浜市泉区,横浜市
-17,横浜市青葉区,横浜市
-18,横浜市都筑区,横浜市
-19,川崎市川崎区,川崎市
-20,川崎市幸区,川崎市
-21,川崎市中原区,川崎市
-22,川崎市高津区,川崎市
-23,川崎市多摩区,川崎市
-24,川崎市宮前区,川崎市
-25,川崎市麻生区,川崎市
-26,相模原市緑区,県央地区
-27,相模原市中央区,県央地区
-28,相模原市南区,県央地区
-29,横須賀市,横須賀三浦地区
-30,平塚市,湘南地区
-31,鎌倉市,横須賀三浦地区
-32,藤沢市,湘南地区
-33,小田原市,県西地区
-34,茅ヶ崎市,湘南地区
-35,逗子市,横須賀三浦地区
-36,三浦市,横須賀三浦地区
-37,秦野市,湘南地区
-38,厚木市,県央地区
-39,大和市,県央地区
-40,伊勢原市,湘南地区
-41,海老名市,県央地区
-42,座間市,県央地区
-43,南足柄市,県西地区
-44,綾瀬市,県央地区
-45,葉山町,横須賀三浦地区
-46,寒川町,湘南地区
-47,大磯町,湘南地区
-48,二宮町,県西地区
-49,中井町,県西地区
-50,大井町,県西地区
-51,松田町,県西地区
-52,山北町,県西地区
-53,開成町,県西地区
-54,箱根町,県西地区
-55,真鶴町,県西地区
-56,湯河原町,県西地区
-57,愛川町,県央地区
-58,清川村,県央地区
