--- data/tokyo-all/arealist.csv
@@ -1,63 +1,0 @@
-area_id,area_name,area_block
-1,台東区,23区東部
-2,墨田区,23区東部
-3,江東区,23区東部
-4,足立区,23区東部
-5,江戸川区,23区東部
-6,葛飾区,23区東部
-7,荒川区,23区東部
-8,品川区,23区西部
-9,目黒区,23区西部
-10,大田区,23区西部
-11,世田谷区,23区西部
-12,中野区,23区西部
-13,杉並区,23区西部
-14,練馬区,23区西部
-15,板橋区,23区西部
-16,北区,23区西部
-17,千代田区,23区都心部
-18,中央区,23区都心部
-19,港区,23区都心部
-20,新宿区,23区都心部
-21,文京区,23区都心部
-22,渋谷区,23区都心部
-23,豊島区,23区都心部
-24,清瀬市,多摩北部
-25,東大和市,多摩北部
-26,東村山市,多摩北部
-27,東久留米市,多摩北部
-28,小平市,多摩北部
-29,西東京市,多摩北部
-30,国分寺市,多摩北部
-31,小金井市,多摩北部
-32,武蔵野市,多摩北部
-33,武蔵村山市,多摩北部
-34,国立市,多摩南部
-35,三鷹市,多摩南部
-36,調布市,多摩南部
-37,狛江市,多摩南部
-38,稲城市,多摩南部
-39,町田市,多摩南部
-40,多摩市,多摩南部
-41,府中市,多摩南部
-42,日野市,多摩南部
-43,立川市,多摩南部
-44,八王子市,多摩西部
-45,青梅市,多摩西部
-46,福生市,多摩西部
-47,羽村市,多摩西部
-48,あきる野市,多摩西部
-49,瑞穂町,多摩西部
-50,日の出町,多摩西部
-51,奥多摩町,多摩西部
-52,昭島市,多摩西部
-53,大島町,島しょ部
-54,利島村,島しょ部
-55,新島村,島しょ部
-56,神津島村,島しょ部
-57,三宅村,島しょ部
-58,御蔵島村,島しょ部
-59,青ヶ島村,島しょ部
-60,八丈町,島しょ部
-61,檜原村,多摩西部
-62,小笠原村,島しょ部
