+++ public/data/block/syonan.json
@@ -1,1 +1,1 @@
-[{"area_id":37,"name":"峠（バス）","lat":35.358160499,"long":139.1778456485,"status":1,"note":"渋沢丘陵のハイキングに行ったときに本当にあったポスター掲示板"},{"area_id":32,"name":"ラーメン二郎藤沢店","lat":35.344486923,"long":139.4821655211,"status":0,"note":"前を通ったことはあっていつも列ができています"}]
+[]
