+++ public/data/block/yokohama.json
@@ -1,1 +1,1 @@
-[{"area_id":10,"name":"味噌ラーメン 雪ぐに","lat":35.4203216004,"long":139.535722727,"status":0,"note":"ネットで評判の店"},{"area_id":10,"name":"支那そばや 戸塚本店","lat":35.4017969146,"long":139.5317783844,"status":0,"note":"一時代前にラーメン博物館で行列もできていたお店。"},{"area_id":10,"name":"湘南の手前","lat":35.3919455977,"long":139.5211915672,"status":1,"note":"面白いネーミングです"},{"area_id":10,"name":"地鶏豚骨ひなわ","lat":35.3997624644,"long":139.5317518808,"status":0,"note":"新しく発見した店"},{"area_id":10,"name":"中華そば しんの助","lat":35.4010508075,"long":139.5297742114,"status":0,"note":"レビュー見て一度行ったけど・・・うーん"},{"area_id":9,"name":"光輪寺","lat":35.5605249643,"long":139.6359403314,"status":1,"note":"横浜にある国際的な仏教寺院。"},{"area_id":9,"name":"興禅寺","lat":35.5574933289,"long":139.6214015766,"status":1,"note":"「興禅寺の沿革」には、仁寿3年開山と書かれています。西暦853年です。"},{"area_id":9,"name":"本長寺","lat":35.5283536724,"long":139.6376829845,"status":2,"note":"すだれ桜が奇麗なお寺です"},{"area_id":9,"name":"常真寺","lat":35.5338200355,"long":139.6106756004,"status":6,"note":"本堂の前に立つ巨大な榧(かや)の木がとても存在感のあるお寺です。"},{"area_id":5,"name":"手打ち蕎麦 鎌倉 梅の木 南太田店","lat":35.4361044,"long":139.6189123,"status":1,"note":"閉店してるようです"},{"area_id":11,"name":"神勝軒下永谷","lat":35.4082409159,"long":139.5642936225,"status":1,"note":"池袋のつけ麺の店がルーツ。おいしいよ"}]
+[]
