+++ public/data/summary.json
@@ -3,13 +3,13 @@
   "2": 0.0,
   "3": 0.0,
   "4": 0.0,
-  "5": 1.0,
+  "5": 0.0,
   "6": 0.0,
   "7": 0.0,
   "8": 0.0,
-  "9": 1.0,
-  "10": 0.2,
-  "11": 1.0,
+  "9": 0.0,
+  "10": 0.0,
+  "11": 0.0,
   "12": 0.0,
   "13": 0.0,
   "14": 0.0,
@@ -29,13 +29,13 @@
   "28": 0.0,
   "29": 0.0,
   "30": 0.0,
-  "31": 0.6,
+  "31": 0.85614,
   "32": 0.0,
-  "33": 1.0,
+  "33": 0.0,
   "34": 0.0,
   "35": 0.0,
   "36": 0.0,
-  "37": 1.0,
+  "37": 0.0,
   "38": 0.0,
   "39": 0.0,
   "40": 0.0,
@@ -57,5 +57,5 @@
   "56": 0.0,
   "57": 0.0,
   "58": 0.0,
-  "total": 0.63158
+  "total": 0.85614
 }
