+++ public/data/summary_absolute.json
@@ -8,7 +8,7 @@
   "7": "0",
   "8": "0",
   "9": "0",
-  "10": "4",
+  "10": "0",
   "11": "0",
   "12": "0",
   "13": "0",
@@ -29,8 +29,8 @@
   "28": "0",
   "29": "0",
   "30": "0",
-  "31": "2",
-  "32": "1",
+  "31": "41",
+  "32": "0",
   "33": "0",
   "34": "0",
   "35": "0",
@@ -57,5 +57,5 @@
   "56": "0",
   "57": "0",
   "58": "0",
-  "total": "7"
+  "total": "41"
 }
